link to my domain

+++ src/components/MenuTree.js
@@ -16,7 +16,7 @@ export default class MenuTree extends Component {
   }
 
   async deleteMenu(deleteId){
-    const res = await axios.delete(`http://itsmyseatvcserver-env.drc3wmhbci.ap-northeast-2.elasticbeanstalk.com/api/cafes/menu/${deleteId}`);
+    const res = await axios.delete(`https://api.knowgari.com/api/cafes/menu/${deleteId}`);
     if(res.data.status === 'success'){
       window.location.reload();
     } else {
@@ -31,7 +31,7 @@ export default class MenuTree extends Component {
         `메뉴명 : ${this.name.current.value} 가격 : ${price}원이 맞나요?`
       )
     ) {
-      axios.post(`/api/cafes/menu/${window.location.href.slice(34)}`, {
+      axios.post(`https://api.knowgari.com/api/cafes/menu/${window.location.href.slice(34)}`, {
         name: this.name.current.value,
         price: price,
         id: this.state.updateModal.id,
@@ -135,12 +135,12 @@ export default class MenuTree extends Component {
               <div
                 className="delete"
                 onClick={e => {
-                  //if (window.confirm('정말 삭제하시겠습니까?')) {
+                  if (window.confirm('정말 삭제하시겠습니까?')) {
                     this.deleteMenu(
                       e.currentTarget.parentNode.getAttribute('data-id')
                     );
                     alert('삭제되었습니다');
-                  //}
+                  }
                 }}
               >
                 <span>삭제</span>
